+++ src/util/pm.js
@@ -1,4 +1,5 @@
-export function isCollapsed (pm) {
-  const {anchor, head} = pm.selection
-  return (anchor === head)
+export function focusedIndex (pm) {
+  const selection = pm.selection
+  if (!selection || selection.anchor == null) return
+  return pm.doc.childBefore(pm.selection.anchor).index
 }