+++ test/ed.js
@@ -86,6 +86,17 @@ describe('Ed', function () {
       }
       expect(no_container).to.throw('Missing options.container')
     })
+    it('calls options.onMount', function (done) {
+      ed = new Ed(
+        { container: mount
+        , initialContent: []
+        , onChange: function () {}
+        , onShareUrl: function () {}
+        , onShareFile: function () {}
+        , onMount: function () { done() }
+        }
+      )
+    })
   })
 
   describe('Content mounting and merging', function () {
@@ -110,6 +121,7 @@ describe('Ed', function () {
     })
     afterEach(function () {
       ed.teardown()
+      mount.parentNode.removeChild(mount)
     })
 
     it('on mount it has expected editable html structure', function () {
@@ -338,6 +350,7 @@ describe('Ed', function () {
     })
     afterEach(function () {
       ed.teardown()
+      mount.parentNode.removeChild(mount)
     })
 
     describe('Mounting', function () {
@@ -361,6 +374,62 @@ describe('Ed', function () {
           ]
         expect(content).to.deep.equal(expected)
       })
+    })
+  })
+
+  describe('Command interface', function () {
+    let mount, ed
+    const fixture =
+      [ {type: 'h1', html: '<h1>Title</h1>'}
+      , {type: 'text', html: '<p><a href="moo">link</a></p>'}
+      ]
+
+    afterEach(function () {
+      ed.teardown()
+      mount.parentNode.removeChild(mount)
+    })
+
+    it('returns commands on mount', function (done) {
+      mount = document.createElement('div')
+      document.body.appendChild(mount)
+
+      function onCommandsChanged (commands) {
+        expect(commands['heading:make1']).to.equal('disabled')
+        expect(commands['paragraph:make']).to.equal('inactive')
+        done()
+      }
+
+      ed = new Ed(
+        { container: mount
+        , initialContent: fixture
+        , onChange: function () {}
+        , onShareUrl: function () {}
+        , onShareFile: function () {}
+        , onCommandsChanged
+        }
+      )
+    })
+
+    it('correctly executes command', function (done) {
+      mount = document.createElement('div')
+      document.body.appendChild(mount)
+
+      function onMount () {
+        ed.execCommand('paragraph:make')
+        const content = ed.getContent()
+        expect(content[0].type).to.equal('text')
+        done()
+      }
+
+      ed = new Ed(
+        { container: mount
+        , initialContent: fixture
+        , onChange: function () {}
+        , onShareUrl: function () {}
+        , onShareFile: function () {}
+        , onMount
+        }
+      )
     })
   })
 })